Give the named key codes a type of their own
`keycodes/codes.rs` held 184 `pub const K_*: c_int`, the densest run of
constants in the tree. They are now `Key`, a `#[repr(i32)]` enum whose
discriminants are still `extra(KE_FOO)` and `termcap(b'k', b'u')` — the
enum *is* upstream's definition of each code, not a second copy of it.

## The seam

A key and a character share one `c_int`: `vgetc`, the typeahead and every
`cmdchar` carry either, and upstream tells them apart by sign. So no field
became an enum and no layout moved. `Key` is the *named* half of that space,
`Key::try_from` is the way in, and `Key::code()` is the way back out for the
many places that still hold a key as a number. `K_SPECIAL` stays a `const`:
it is the escape byte a key travels as, not a key.

`TryFrom` binary searches `Key::ALL`, which is sorted by code and carries a
`const` assertion that it still is, so adding a key is a matter of putting
it in the right place.

## Dispatches

Every `match c { K_UP => … }` now matches `Key::try_from(c)`. Where the arms
mix keys with character constants — most of `edit/key.rs`,
`ex_getln/handlekey.rs`, `window/cmd.rs`, `terminal/` — the character half
is a pattern again because `NotAKey` carries the number:

    Ok(Key::Home | Key::Khome | Key::SHome | Key::CHome) | Err(NotAKey(Ctrl_B)) => …

which keeps upstream's alternations intact instead of splitting a body per
half. Nothing in an `Err(NotAKey(…))` arm here can be negative, so no key
value can hide behind one.

`normal/dispatch.rs` loses eight local `const K_C_LEFT: c_int = -(253 + …)`
definitions: transpiler residue that inlined the same value the module
constant already had.

## Elsewhere

`scripts/gen-keycodes-lua.py` reads the key-name table's first column, and
`crates/nvim/src/keycodes.lua` still comes out byte for byte the same: a
variant name is the C name with each underscore-separated word capitalised
(`K_S_F1` is `Key::SF1`), which inverts by splitting on the capitals again.

`nfa_get_reganch` answers `bool` now rather than the 0/1 `c_int` it always
meant, which pays for the one `-> c_int` `Key::code()` adds.

`const_c_int` 778 -> 594.

+++ crates/nvim/src/regexp/nfa/exec.rs
@@ -349,7 +349,7 @@ pub(crate) unsafe fn nfa_regcomp(expr: *mut uint8_t, re_flags: c_int) -> *mut re
             unsafe { (*prog).has_backref = rex.nfa_has_backref() };
             unsafe { (*prog).nsubexp = regnpar.get() };
             nfa_postprocess(prog);
-            unsafe { (*prog).reganch = nfa_get_reganch((*prog).start, 0) };
+            unsafe { (*prog).reganch = c_int::from(nfa_get_reganch((*prog).start, 0)) };
             unsafe { (*prog).regstart = nfa_get_regstart((*prog).start, 0) };
             unsafe { (*prog).match_text = nfa_get_match_text((*prog).start) };
             unsafe { (*prog).reghasz = re_has_z.get() };
